Updates to data processing and UI
Updated the data processing to filter the data to try to find only sites that have water discharge info available and using those as the choices for the drop-down menu to select a site. Made updates to the UI to acomodate this. Now user enters a state and the name of the river as a search term then clicks a button to start the data processing.